Verify that a custom builder is used everywhere
This commit changes the error builder function to make something different than
just returning a string. This way, after updating all assertions against errors
in tests, I've verified that the builder is used, and that the error code is
never lost when transforming single errors into lists and dictionaries of those.